+++ airbyte-integrations/connectors/source-relational-db/src/main/java/io/airbyte/integrations/source/relationaldb/CursorInfo.java
@@ -67,7 +67,7 @@ public class CursorInfo {
     return this;
   }
 
-  public CursorInfo setCursorRecordCount(final int cursorRecordCount) {
+  public CursorInfo setCursorRecordCount(final long cursorRecordCount) {
     this.cursorRecordCount = cursorRecordCount;
     return this;
   }
