+++ airbyte-integrations/connectors/source-relational-db/src/test/java/io/airbyte/integrations/source/relationaldb/state/StateTestConstants.java
@@ -28,7 +28,7 @@ public final class StateTestConstants {
   public static final String CURSOR_FIELD1 = "year";
   public static final String CURSOR_FIELD2 = "generation";
   public static final String CURSOR = "2000";
-  public static final int CURSOR_RECORD_COUNT = 19;
+  public static final long CURSOR_RECORD_COUNT = 19L;
 
   private StateTestConstants() {}
 
@@ -39,12 +39,12 @@ public final class StateTestConstants {
         .withCursor(cursor));
   }
 
-  public static Optional<DbStreamState> getState(final String cursorField, final String cursor, final int cursorRecordCount) {
+  public static Optional<DbStreamState> getState(final String cursorField, final String cursor, final long cursorRecordCount) {
     return Optional.of(new DbStreamState()
         .withStreamName(STREAM_NAME1)
         .withCursorField(Lists.newArrayList(cursorField))
         .withCursor(cursor)
-        .withCursorRecordCount((long) cursorRecordCount));
+        .withCursorRecordCount(cursorRecordCount));
   }
 
   public static Optional<ConfiguredAirbyteCatalog> getCatalog(final String cursorField) {
